fix(agent): preflight resource policy before allocating stores
