+++ packages/agent/src/sync/catchup-pass-policy.ts
@@ -519,15 +519,12 @@ export function resolveSwmCatchupMaxPasses(raw: string | undefined, onRejected?:
  * explicit value keeps parsing pure and lets each driver inject the same shape
  * into its continuation loop.
  */
-export function resolveSwmCatchupPassConfig(environment: {
-  DKG_SWM_CATCHUP_PASS_BUDGET_MS?: string;
-  DKG_SWM_CATCHUP_MAX_PASSES?: string;
-} = {
+export function resolveSwmCatchupPassConfig(environment: Readonly<Record<string, string | undefined>> = {
   DKG_SWM_CATCHUP_PASS_BUDGET_MS: process.env.DKG_SWM_CATCHUP_PASS_BUDGET_MS,
   DKG_SWM_CATCHUP_MAX_PASSES: process.env.DKG_SWM_CATCHUP_MAX_PASSES,
-}, onRejected?: RejectedResourceSetting): CatchupPassConfig {
-  return {
+}, onRejected?: RejectedResourceSetting): Readonly<CatchupPassConfig> {
+  return Object.freeze({
     budgetMs: resolveSwmCatchupPassBudgetMs(environment.DKG_SWM_CATCHUP_PASS_BUDGET_MS, onRejected),
     maxPasses: resolveSwmCatchupMaxPasses(environment.DKG_SWM_CATCHUP_MAX_PASSES, onRejected),
-  };
+  });
 }
